feat: update tsconfig
