+++ src/components/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { default as Header } from './Header';
+export { default as Card } from './cards/Card';
