+++ tsconfig.app.json
@@ -20,7 +20,16 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true
+    "noUncheckedSideEffectImports": true,
+    "paths": {
+      "@api": ["./src/api/index.ts"],
+      "@assets/*": ["./src/assets/*"],
+      "@components": ["./src/components/*", "./src/components"],
+      "@layout": ["./src/layout"],
+      "@pages": ["./src/pages"],
+      "@types": ["./src/types"],
+      "@utils": ["./src/utils"]
+    }
   },
   "include": ["src", "tailwind.config.js"]
 }
